Use alternative address in GetSpecializedLightPointer() to prevent shadow flicker
- Add code to Reset() to release silhouette surface/texture (thanks Elisha)

+++ Patches/Patches.h
@@ -14,7 +14,7 @@ void *GetCutscenePosPointer();
 BYTE *GetFlashLightRenderPointer();
 void *GetJamesPosPointer();
 void *GetRoomIDPointer();
-BYTE *GetSpecializedLightPointer();
+DWORD *GetSpecializedLightPointer();
 
 // Function forward declaration
 void DisableCDCheck();
